third_party/blink/web_tests/http/tests/security/xss-DENIED-window-name-alert-expected.txt

CONSOLE MESSAGE: line 9: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
